+++ aladdin-lamp/src/main/java/com/yeem/lamp/domain/objvalue/Plan.java
@@ -8,7 +8,7 @@ import java.math.BigDecimal;
 public class Plan {
     private PlanType planType;
     private BigDecimal price;
-    private BigDecimal dataTraffic;
+    private Integer dataTraffic;
     private Integer period;
     private String title;
     private String introduce;
